Refactor IT classes and packages
