+++ integration-test/src/test/java/org/apache/iotdb/db/it/utils/AlignedWriteUtil.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.it.aligned;
+package org.apache.iotdb.db.it.utils;
 
 import org.apache.iotdb.it.env.EnvFactory;
